Finished migrating the models, views, urls and forms corresponding to the orders app
